==> 0. Updating The Dataset.ipynb <==
# Librerias pertinentes para la recoleccion de datos

import yfinance
# Libreria para el desahabilitacion de advertencias despues del debug cell
import warnings 
warnings.filterwarnings("ignore")
import pandas as pd

raw_data=yfinance.download(tickers='^FTSE ^GDAXI ^GSPC ^N225',start='1994-01-07',end='2023-01-12',group_by='ticker',interval='1d',auto_adjust=True)


## Funcion Download
* <b> Download </b>: Funcion que permite extraer los datos de yahoo finance
* <b> Tickers </b>: Nombre o id de los activos financieros
* <b> Start </b>: Inicio de tiempo de recoleccion de datos
* <b> End </b>: Fin de tiempo de recoleccion de datos
* <b> Groupby </b>: Funcion que permite agrupar mediante un indice como tickers
* <b> Interval </b>:Distancia de tiempo entre 2 observaciones de datos,se extraera datos de tiempo segun su frecuencia como 1d-un dia


# Copia de datos a otra variable para remover o alterar datos o indices
df_comp = raw_data.copy()

# Adding new columns to the data set
df_comp['spx'] = df_comp['^GSPC'].Close
df_comp['dax'] = df_comp['^GDAXI'].Close
df_comp['ftse'] = df_comp['^FTSE'].Close
df_comp['nikkei'] = df_comp['^N225'].Close

#Se eliminan los datos de las indices de la primera fila
#por motivos de cierre de bolsa en otras zonas horarias
df_comp=df_comp.iloc[1:]
#Se eliminan los datos de los tickers originales por remover datos redudantes
f=['^FTSE','^GDAXI','^GSPC','^N225']
for a in f:
    del df_comp[a]
df_comp=df_comp.asfreq('b') # Corregir frecuencia de datos como b-Bimestral
# Imputacion de datos sobre variables con datos nulos
df_comp.fillna(method='ffill',inplace=True)

print (df_comp.head()) #Muestra los datos de las primeras columnas para verificar la correcta extracion de datos
print (df_comp.tail()) #Autotenticacion de datos al mostrar los ultimas filas de los mismo para asegurar de que la serie esta correctamente extraida hasta el ultimo dia de negociacion

new_data=yfinance.download(tickers='^GSPC ^FTSE ^GDAXI ^N225',start='2018-01-29',end='2023-10-11',group_by='ticker',auto_adjust=True,interval='1d')
new_data.round(2)

df_copy = new_data.copy()
df_copy['spx'] = df_copy['^GSPC'].Close
df_copy['dax'] = df_copy['^GDAXI'].Close
df_copy['ftse'] = df_copy['^FTSE'].Close
df_copy['nikkei'] = df_copy['^N225'].Close

#Se eliminan los datos de las indices de la primera fila
#por motivos de cierre de bolsa en otras zonas horarias
df_copy=df_copy.iloc[1:]
#Se eliminan los datos de los tickers originales por remover datos redudantes
f=['^FTSE','^GDAXI','^GSPC','^N225']
for a in f:
    del df_copy[a]
df_comp=df_copy.asfreq('b') # Corregir frecuencia de datos como b-Bimestral
# Imputacion de datos sobre variables con datos nulos
df_copy.fillna(method='ffill',inplace=True)

df_copy.to_csv('Index_2018_new.csv')

df_copy

# Lee los dos archivos CSV en DataFrames separados
df1 = pd.read_csv('Index2018.csv')
df2 = pd.read_csv('Index_2018_new.csv')

# Fusiona los DataFrames en uno solo
Index_united= pd.concat([df1, df2], ignore_index=True)

# Ahora `merged_df` contiene la fusión de los dos archivos CSV
Index_united

df2=df2.iloc[1:]

df1

df2.rename(columns={'Unnamed: 0':'date'},inplace=True)

df1,df2

Index_united=pd.concat([df1,df2],axis=0,ignore_index=True)
Index_united

Index_united.to_csv('Index_united.csv')
