==> 2. Creating a Time Series Object in Python.ipynb <==
### Importacion de Librerias

import pandas as pd
import numpy as np

### Importacion de los datos

raw_csv_data = pd.read_csv("Index_united.csv") 
df_comp=raw_csv_data.copy() 

### Limpieza de los datos

#Muestra de las columnas del dataframe
df_comp.columns
df_comp.drop(columns='Unnamed: 0',inplace=True)


#### Conversion de los datos de la columna **date** en formato fecha

from datetime import datetime

date_string = "13/01/1994"
format = "%d/%m/%Y"

try:
    date = datetime.strptime(date_string, format)
    print(date)
except ValueError as e:
    print("Error:", e)


### Tamaño del tiempo de la data

df_comp.describe()## Observacion de los momentos muestrales de cada serie de los precios de cada columna

df_comp.head()##Primeros registros de la data

df_comp.date.describe()  ##Momentos muestrales de la serie date

df_comp.head()

df_comp.date.describe()

### Conversion de la columna object a Datetime

df_comp.date=pd.to_datetime(df_comp.date,dayfirst=True)

df_comp.info()

### Marcando el indice de cada cotizacion

df_comp.set_index("date")

df_comp.set_index("date", inplace=True) #Sitiando sobre la misma variable

df_comp.head()

#### Base de datos guardada

df_comp.to_csv('Index_modified.csv')

### Muestra de frecuencia preferida

df_compd=df_comp.asfreq('d') ##Freecuencia diaria

df_compd.shape

df_compb=df_comp.asfreq('b') # Frecuencia bimestral

df_compb.head() ## Primeros registros

### Handling Missing Values

df_compb.isna() ## Verificaion de nulos en el dataframe

df_compb.isna().sum() ##Suma de los datos nulos del dataframe de cada columna

df_compb

##Imputacion  de datos sobre los datos nulos de manera que reemplazará los valores faltantes con el valor de la fila anterior en la misma columna
df_comp.spx=df_comp.spx.fillna(method='ffill')

df_comp.isna().sum()#Suma de datos nulos

##Imputacion  de datos sobre los datos nulos de manera queque reemplazará los valores faltantes con el valor de la fila siguiente en la misma columna.
df_comp.ftse=df_comp.ftse.fillna(method='bfill')

##Imputacion  de datos sobre los datos nulos de manera queque reemplazará los valores faltantes con el valor de la media
df_comp.dax=df_comp.dax.fillna(value = df_comp.dax.mean())

### Simplificando los datos

## Nuevaa Variable **market_value** (capitakizacion de mercado) o activo corresponde al indice **GSPC**
df_comp['market_value']=df_comp.spx 

df_comp.describe()

del df_comp['spx'] # Eliminacion original de los datos de GSPC

df_comp.describe() # Descripcion de los datos

# Eliminacion original de los datos de cada indice
del df_comp['dax'] 
del df_comp['ftse']
del df_comp['nikkei']

df_comp.describe()

### Separacion de la data en train y test 

size = int(len(df_comp)*0.8) #Separacion de data original en d_test y df = df_train donde se podra pobrar la eficacia de los modelos para entrenar y predecir nuevos datos

df = df_comp.iloc[:size]

df_test = df_comp.iloc[size:]

df.tail()

df_test.head()
